+++ spec/guard/commands/show_spec.rb
@@ -4,8 +4,7 @@ describe 'Guard::Interactor::SHOW' do
 
   describe '#perform' do
     it 'outputs the DSL description' do
-      ::Guard::DslDescriber.should_receive(:show).and_return 'SHOW OUTPUT'
-      Pry.output.should_receive(:puts).with 'SHOW OUTPUT'
+      ::Guard::DslDescriber.should_receive(:show)
       Pry.run_command 'show'
     end
   end