+++ spec/guard/dsl_describer_spec.rb
@@ -1,6 +1,7 @@
 # encoding: utf-8
 
 require 'spec_helper'
+require 'formatador'
 
 describe Guard::DslDescriber do
 
@@ -27,21 +28,30 @@ describe Guard::DslDescriber do
   before do
     stub_const 'Guard::Test', Class.new(Guard::Guard)
     stub_const 'Guard::Another', Class.new(Guard::Guard)
+
+    @output = ''
+
+    # Strip escape sequences
+    STDOUT.stub(:tty?).and_return false
+
+    # Capture formatador output
+    Thread.current[:formatador] = Formatador.new
+    Thread.current[:formatador].stub(:print) do |msg|
+      @output << msg
+    end
   end
 
   describe '.list' do
-    let(:output) do
-      <<-OUTPUT.gsub(/^\s+|\n$/, '')
-        +------------+--------------+
-        |  Available Guard plugins  |
-        +------------+--------------+
-        | Plugin     | In Guardfile |
-        +------------+--------------+
-        | Another    | ✔            |
-        | Even       | ✘            |
-        | More       | ✘            |
-        | Test       | ✔            |
-        +------------+--------------+
+    let(:result) do
+      <<-OUTPUT
+  +---------+-----------+
+  | Plugin  | Guardfile |
+  +---------+-----------+
+  | Another | ✔         |
+  | Even    | ✘         |
+  | More    | ✘         |
+  | Test    | ✔         |
+  +---------+-----------+
       OUTPUT
     end
 
@@ -50,30 +60,31 @@ describe Guard::DslDescriber do
     end
 
     it 'lists the available Guards when they\'re declared as strings or symbols' do
-      # Delete is needed for 1.8.7 compatibility only
-      ::Guard::DslDescriber.list(:guardfile_contents => guardfile).to_s.delete(' ').should eq(output.delete(' '))
+      ::Guard::DslDescriber.list(:guardfile_contents => guardfile)
+      @output.should eql(result)
     end
   end
 
   describe '.show' do
-    let(:output) do
-      <<-OUTPUT.gsub(/^\s+|\n$/, '')
-       +---------+---------+--------+-------+
-       |        Guardfile structure         |
-       +---------+---------+--------+-------+
-       | Group   | Plugin  | Option | Value |
-       +---------+---------+--------+-------+
-       | Default | Test    | a      | :b    |
-       |         |         | c      | :d    |
-       | A       | Test    | x      | 1     |
-       |         |         | y      | 2     |
-       | B       | Another |        |       |
-       +---------+---------+--------+-------+
+    let(:result) do
+      <<-OUTPUT
+  +---------+---------+--------+-------+
+  | Group   | Plugin  | Option | Value |
+  +---------+---------+--------+-------+
+  | Default | Test    | a      | :b    |
+  |         |         | c      | :d    |
+  +---------+---------+--------+-------+
+  | A       | Test    | x      | 1     |
+  |         |         | y      | 2     |
+  +---------+---------+--------+-------+
+  | B       | Another |        |       |
+  +---------+---------+--------+-------+
       OUTPUT
     end
 
     it 'shows the Guards and their options' do
-      ::Guard::DslDescriber.show(:guardfile_contents => guardfile).to_s.should eq(output)
+      ::Guard::DslDescriber.show(:guardfile_contents => guardfile)
+      @output.should eql(result)
     end
   end